# gen/docs/swagger/swagger.yaml
definitions:
  domain.UserStatus:
    enum:
    - pending
    - active
    - locked
    - disabled
    type: string
    x-enum-varnames:
    - UserStatusPending
    - UserStatusActive
    - UserStatusLocked
    - UserStatusDisabled
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
      trace_id:
        type: string
    type: object
...
  handlers.SessionBulkRevokeResponse:
    properties:
      message:
        type: string
      revoked_count:
        type: integer
      tokens_revoked:
        type: integer
    type: object
  handlers.SessionListResponse:
    properties:
      sessions:
        items:
          $ref: '#/definitions/handlers.SessionPayload'
        type: array
      total:
        type: integer
    type: object
  handlers.SessionPayload:
    properties:
      created_at:
        type: string
      device_id:
        type: string
      device_label:
        type: string
      expires_at:
        type: string
      id:
        type: string
      ip_first:
        type: string
      ip_last:
        type: string
      is_active:
        type: boolean
      is_current:
        type: boolean
      last_seen:
        type: string
      refresh_token_id:
        type: string
      revoke_reason:
        type: string
      revoked_at:
        type: string
      user_agent:
        type: string
      user_id:
        type: string
    type: object
...
  handlers.SessionValidateRequest:
    properties:
      session_id:
        type: string
    required:
    - session_id
    type: object
  handlers.SessionValidateResponse:
    properties:
      session:
        $ref: '#/definitions/handlers.SessionPayload'
      valid:
        type: boolean
    type: object
  handlers.TokenRefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handlers.TokenRefreshResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/handlers.UserSummary'
    type: object
  handlers.UserSummary:
    properties:
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      roles:
        items:
          type: string
        type: array
      status:
        $ref: '#/definitions/domain.UserStatus'
      username:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/auth/token/refresh:
    post:
      consumes:
      - application/json
      description: Issues a new access token using a valid refresh token.
      parameters:
      - description: Refresh request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TokenRefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TokenRefreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Refresh an access token
      tags:
      - Authentication
...
  /api/v1/sessions:
    delete:
      description: Revokes all active sessions for the authenticated user (including
        the current one).
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Must be true to confirm bulk revocation
        in: query
        name: all
        required: true
        type: boolean
      - description: Optional revocation reason
        in: query
        name: reason
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SessionBulkRevokeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Revoke all sessions
      tags:
      - Sessions
    get:
      description: Retrieves sessions for the authenticated user with optional filtering.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: When true (default) only active sessions are returned
        in: query
        name: active_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SessionListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: List sessions for authenticated user
      tags:
      - Sessions
  /api/v1/sessions/{session_id}:
    delete:
      description: Revokes an active session owned by the authenticated user.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Session identifier
        in: path
        name: session_id
        required: true
        type: string
      - description: Optional revocation reason
        in: query
        name: reason
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Session revoked
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Revoke a specific session
      tags:
      - Sessions
  /api/v1/sessions/others:
    delete:
      description: Revokes all active sessions except the current session.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Optional revocation reason
        in: query
        name: reason
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SessionBulkRevokeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Revoke all other sessions
      tags:
      - Sessions
  /api/v1/sessions/validate:
    post:
      consumes:
      - application/json
      description: Checks whether the provided session is still valid.
      parameters:
      - description: Session validation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SessionValidateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SessionValidateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Validate a session
      tags:
      - Sessions
...
swagger: "2.0"